Improve service type db

+++ database/migrations/2022_08_16_090618_create_service_type_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 128);
             $table->text('description');
-            $table->unsignedSmallInteger('cost');
+            $table->unsignedInteger('cost');
             $table->timestamps();
         });
     }
